requirements.txt
streamlit
tensorflow
Pillow
pandas
kaggle
